this commit fixes #1088 fixes #1085 fixes #1081 fixes #1073 fixes #1028

+++ app/src/main/java/com/fastaccess/ui/modules/repos/issues/create/CreateIssuePresenter.java
@@ -79,7 +79,7 @@ public class CreateIssuePresenter extends BasePresenter<CreateIssueMvp.View> imp
                         createIssue.setAssignees(Stream.of(users).map(User::getLogin).collect(Collectors.toCollection(ArrayList::new)));
                     }
                     if (milestoneModel != null) {
-                        createIssue.setMilestone(milestoneModel.getNumber());
+                        createIssue.setMilestone((long) milestoneModel.getNumber());
                     }
                 }
                 makeRestCall(RestProvider.getIssueService(isEnterprise()).createIssue(login, repo, createIssue),
